+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add support for `serde_json::Map` (https://github.com/fiberplane/fp-bindgen/pull/163)
 - Replace the `rust_plugin_module` and `rust_wasmer_runtime_module` annotations with a single `rust_module` annotation.
 - Migrate to Wasmer 3 (https://github.com/fiberplane/fp-bindgen/pull/157)
+- Struct fields annotated with `skip_serializing_if` are treated as optional by the Typescript binding generator.
 
 ## 2.4.0